# docs/architecture/openapi.json
{
  "components": {
    "schemas": {
      "AcceptanceReportResponse": {
        "properties": {
          "canonical_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Asset Id"
          },
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "claim_status": {
            "enum": [
              "observed",
              "submitted",
              "materialization_pending",
              "under_review",
              "accepted",
              "accepted_duplicate",
              "rejected",
              "expired"
            ],
            "title": "Claim Status",
            "type": "string"
          },
          "decided_at": {
            "title": "Decided At",
            "type": "string"
          },
          "decision_reason": {
            "title": "Decision Reason",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "status": {
            "enum": [
              "accepted",
              "accepted_duplicate",
              "rejected_invalid",
              "rejected_policy",
              "deferred_materialization",
              "deferred_pending_review"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "status",
          "claim_status",
          "decision_reason",
          "decided_at"
        ],
        "title": "AcceptanceReportResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeClaimRequest": {
        "properties": {
          "advertised_source_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Advertised Source Kinds",
            "type": "array"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "ephemeral": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ephemeral"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "unknown",
                  "healthy",
                  "degraded",
                  "online",
                  "offline"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "title": "NodeClaimRequest",
        "type": "object"
      },
      "NodeRecordPublic": {
        "properties": {
          "advertised_source_kinds": {
            "items": {
              "type": "string"
            },
            "title": "Advertised Source Kinds",
            "type": "array"
          },
          "auth_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Auth Scopes",
            "type": "array"
          },
          "auth_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Type"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "ephemeral": {
            "default": false,
            "title": "Ephemeral",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "status": {
            "default": "unknown",
            "enum": [
              "unknown",
              "healthy",
              "degraded",
              "online",
              "offline"
            ],
            "title": "Status",
            "type": "string"
          },
          "token_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Preview"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "node_id",
          "label"
        ],
        "title": "NodeRecordPublic",
        "type": "object"
      },
...
      "SubmitAssetClaimRequest": {
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "kind": {
            "enum": [
              "file",
              "stream",
              "capture-session",
              "compose-output",
              "virtual"
            ],
            "title": "Kind",
            "type": "string"
          },
          "local_ref": {
            "minLength": 1,
            "title": "Local Ref",
            "type": "string"
          },
          "materialization_mode": {
            "enum": [
              "upload",
              "shared-path",
              "pull",
              "stream-ref"
            ],
            "title": "Materialization Mode",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "node_id": {
            "minLength": 1,
            "title": "Node Id",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "source_name": {
            "minLength": 1,
            "title": "Source Name",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "source_name",
          "kind",
          "local_ref",
          "materialization_mode"
        ],
        "title": "SubmitAssetClaimRequest",
        "type": "object"
      },
      "SubmitAssetClaimResponse": {
        "properties": {
          "acceptance": {
            "$ref": "#/components/schemas/AcceptanceReportResponse"
          },
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fingerprint"
          },
          "kind": {
            "enum": [
              "file",
              "stream",
              "capture-session",
              "compose-output",
              "virtual"
            ],
            "title": "Kind",
            "type": "string"
          },
          "local_ref": {
            "title": "Local Ref",
            "type": "string"
          },
          "materialization_mode": {
            "enum": [
              "upload",
              "shared-path",
              "pull",
              "stream-ref"
            ],
            "title": "Materialization Mode",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata",
            "type": "object"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "observed",
              "submitted",
              "materialization_pending",
              "under_review",
              "accepted",
              "accepted_duplicate",
              "rejected",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "status",
          "node_id",
          "source_name",
          "kind",
          "local_ref",
          "materialization_mode",
          "created_at",
          "updated_at",
          "acceptance"
        ],
        "title": "SubmitAssetClaimResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "media-sync-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ingest/claims": {
      "get": {
        "operationId": "list_ingest_claims_api_ingest_claims_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "title": "Response List Ingest Claims Api Ingest Claims Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Ingest Claims",
        "tags": [
          "ingest"
        ]
      },
      "post": {
        "operationId": "submit_ingest_claim_api_ingest_claims_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-media-sync-node-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Media-Sync-Node-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAssetClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAssetClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Ingest Claim",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/ingest/claims/prune": {
      "post": {
        "operationId": "prune_ingest_claims_api_ingest_claims_prune_post",
        "parameters": [
          {
            "in": "query",
            "name": "older_than_seconds",
            "required": false,
            "schema": {
              "default": 86400,
              "minimum": 1,
              "title": "Older Than Seconds",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Prune Ingest Claims Api Ingest Claims Prune Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prune Ingest Claims",
        "tags": [
          "ingest"
        ]
      }
    },
    "/api/ingest/claims/{claim_id}": {
      "delete": {
        "operationId": "delete_ingest_claim_api_ingest_claims__claim_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Delete Ingest Claim Api Ingest Claims  Claim Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ingest Claim",
        "tags": [
          "ingest"
        ]
      },
      "get": {
        "operationId": "get_ingest_claim_api_ingest_claims__claim_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Ingest Claim Api Ingest Claims  Claim Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingest Claim",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/api/nodes/{node_id}/claim": {
      "post": {
        "operationId": "claim_node_api_nodes__node_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeRecordPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Node",
        "tags": [
          "nodes"
        ]
      }
    },
...
  }
}